ref: Remove EMS\CommonBundle\Common\Standard\Type
